requirements.txt
numpy==1.21.0  
Pillow==10.4.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
qrcode[pil]==7.4.2
